requirements.txt
certifi==2020.6.20
chardet==3.0.4
dnspython==1.16.0
eventlet==0.29.1
greenlet==0.4.17
idna==2.10
python-engineio==3.13.2
python-socketio==4.6.0
requests==2.24.0
six==1.15.0
urllib3==1.25.11
websocket-client==0.57.0
